# awesomeProject/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CRUD-to-DB-in-GO API",
    "description": "API for managing blog entries with user authentication",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Development server"
    }
  ],
  "paths": {
    "/login": {
      "post": {
        "summary": "User Login",
        "operationId": "loginUser",
        "tags": ["auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/items/all": {
      "get": {
        "summary": "Get all items",
        "operationId": "getAllItems",
        "tags": ["items"],
        "security": [{ "Auth": [] }],
        "responses": {
          "200": {
            "description": "Successful retrieval of items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Item"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "summary": "Delete all items (admin only)",
        "operationId": "deleteAllItems",
        "tags": ["items"],
        "security": [{ "Auth": [] }],
        "responses": {
          "204": {
            "description": "All items deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/items/single/{id}": {
      "get": {
        "summary": "Get an item by ID",
        "operationId": "getItemById",
        "tags": ["items"],
        "security": [{ "Auth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of item",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Item not found"
          }
        }
      }
    },
    "/items/create": {
      "post": {
        "summary": "Create a new item",
        "operationId": "createItem",
        "tags": ["items"],
        "security": [{ "Auth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Item created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/items/update/{id}": {
      "put": {
        "summary": "Update an item",
        "operationId": "updateItem",
        "tags": ["items"],
        "security": [{ "Auth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Item not found"
          }
        }
      }
    },
    "/items/delete/{id}": {
      "delete": {
        "summary": "Delete an item",
        "operationId": "deleteItem",
        "tags": ["items"],
        "security": [{ "Auth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Item deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Item not found"
          }
        }
      }
    },
    "/items/prefetch": {
      "get": {
        "summary": "Prefetch items",
        "operationId": "prefetchItems",
        "tags": ["items"],
        "security": [{ "Auth": [] }],
        "responses": {
          "200": {
            "description": "Successful prefetch of items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Item"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/items/prefetch/all": {
      "get": {
        "summary": "Prefetch all items (admin only)",
        "operationId": "prefetchAllItems",
        "tags": ["items"],
        "security": [{ "Auth": [] }],
        "responses": {
          "200": {
            "description": "Successful prefetch of all items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Item"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/items/admin/update/{id}": {
      "put": {
        "summary": "Update an item (admin only)",
        "operationId": "updateItemAdmin",
        "tags": ["items"],
        "security": [{ "Auth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Item not found"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "isAdmin": {
            "type": "boolean"
          },
          "AuthToken": {
            "type": "string"
          }
        }
      },
      "Item": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "blogname": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "userID": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "ItemCreate": {
        "type": "object",
        "required": ["blogname", "author", "content"],
        "properties": {
          "blogname": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "content": {
            "type": "string"
          }
        }
      },
      "ItemUpdate": {
        "type": "object",
        "properties": {
          "blogname": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "content": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "Auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}